Shaping implemented
When defining stims in netParams, it is possible to include a structure
called 'shape', for example...
'shape': {'switchOnOff': [100, 700, 1300, 1900], 'pulseType':
'gaussian', 'pulsePeriod': 100, 'pulseWidth': 5}
If any/all attributes are missing, the stimulus exists for the entire
duration of the stimulation, shaped as a square wave, with pulse period
and pulse width of 100 ms. This is effectively equivalent to no shaping.