FInitializeHandler should set v with type 0 handler
The FInitializeHandler to set v should be of type 0:
https://www.neuron.yale.edu/neuron/static/new_doc/simctrl/programmatic.html?highlight=finitialize#FInitializeHandler

This means the ion channel variables will be initialised with this value inside the INITIAL block, as opposed to the default value (-65mV). This will alter the behaviour of existing  HH cell simulations (if they don't start at -65) but just for first few ms, and the sim will be more accurate.