Merge list comprehension into loop
Method divConn() contains a list comprehension that loops though a dictionary once to construct a dict subset, then only uses this subset within a follow-up loop to construct connections.
Basically, the subset construction is unnecessary if every element of the dictionary is going to be visited once and only once anyway, so the if clauses that determine the subset have been moved into the follow-up loop.
With PD_NEURON6.py at 0.05 scale, this change reduces cell connection time from ~150 s to ~50 s.