Make post cell sample a dict
In the divConn method, a sampling of post-cells is constructed via list comprehension and frequently calling keys() for a dict.
First up, the keys() method is now called only once outside the preCell loop, while sample construction is done as a dict so that the if check in the following loop is done instantly rather than searching through a list.
Performance can perhaps be improved by just looping through postCellsSample...