Redundant list comprehension
RNG seeding is done by creating an args list in _addCellConn and checking for the appearance of key 'rand'.
This is redundant, as there is a 'for' loop next up that can run through the check anyway.
However, this modification leads only to a small performance boost, which implies that the obvious improvements for this framework are becoming more limited.