Items iterator test
This was a commit just to test whether getting rid of items(), as was done as one of the edits in the previous commit, had any performance effect.
This test was done with the divergent case of PD_NEURON6.py.
It turns out that there is no difference, only because NetPyne recently switched to Python 3. The items() method used to produce an expensive new list in Python 2, but now generates an iterator which is much faster.
So, keeping the items() for Python 3 ends up being an aesthetic choice.