Update netrxd.py
Following a discussion in the rxd meeting, we thought netpyne shouldn't create regions with no sections, (something that happened when looking at an MPI rxd bug and increasing the number of nodes beyond the number of cells). I think something like this might prevent it.
I also added the membrane_flux keyword argument for multicompartment reactions.